apps/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.3
python-multipart==0.0.17
pandas==2.2.3
python-dotenv==1.0.1
pydantic==2.10.3
openai==1.54.4 